requirements.txt
fastapi==0.97.0
Flask==2.2.3
Flask_Cors==3.0.10
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
Requests==2.31.0
tensorflow==2.12.0
uvicorn==0.22.0
